Updated logidze, applicaion controller to set logidze responsible to transfer action as well, moved log methods from project namespace to history concern as it will be used in models, and updated retrieval of log data and formatting into a hash
